RAPRITHPRA/sunita willams/sunita williams/AI-CHATBOT/requirements.txt
# No external dependencies required
# This package uses only the Python standard library 

google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
Pillow>=10.0.0
gtts>=2.3.2
googletrans-py>=4.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
protobuf>=4.23.4
fpdf>=1.7.2
PyPDF2>=3.0.1
pytesseract>=0.3.10
matplotlib>=3.7.2
langchain>=0.0.267
langchain-google-genai>=0.0.1
numpy>=1.24.3
pandas>=2.0.3
pytz>=2023.3
colorama>=0.4.6
tqdm>=4.65.0
nltk>=3.8.1
scikit-learn>=1.3.0
deep-translator>=1.11.4
tensorflow>=2.8.0
opencv-python>=4.6.0 